+++ tests/test_flickr_url_parser.py
@@ -107,6 +107,8 @@ def test_it_can_parse_urls_even_if_the_host_is_a_bit_unusual(url: str) -> None:
         ("farm1.static.flickr.com/82/241708183_dd0847d5c7_o.jpg", "241708183"),
         ("http://flickr.com/photo/17277074@N00/2619974961", "2619974961"),
         ("https://www.flickr.com/photo_zoom.gne?id=196155401&size=m", "196155401"),
+        ("https://www.flickr.com/photos/gracewong/196155401/meta/", "196155401"),
+        ("https://www.flickr.com/photo_exif.gne?id=1427904898", "1427904898"),
     ],
 )
 def test_it_parses_a_single_photo(url: str, photo_id: str) -> None: