+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Add support for parsing several more varieties of URL, based on those seen in th
 
 *   Old-style photo URLs that use `/photo/` instead of `/photos/`, e.g. `http://flickr.com/photo/17277074@N00/2619974961`
 *   Photo URLs that use the `/photo_zoom.gne` path, e.g. `https://www.flickr.com/photo_zoom.gne?id=196155401&size=m`
+*   Photo URLs that use the `/photo_exif.gne` path, e.g. `https://www.flickr.com/photo_exif.gne?id=1427904898`
 
 ## v1.5.2 - 2023-12-12